requirements.txt
gradio
diffusers
transformers
safetensors
accelerate
compel
pillow
tqdm
